chore: rename ScreenSize package name
