+++ modules/gallery/src/jsMain/kotlin/utils/ScreenSize.js.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.compose.storytale.gallery.compose
+package org.jetbrains.compose.storytale.gallery.utils
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalWindowInfo
